Squashed commit of the following:
commit 8a89d432f02fa67e7c00f990f234a60b448788f1
Author: Pedro Rodrigues <prodrigues@singlestore.com>
Date:   Tue Jul 29 02:37:20 2025 +0100

    improve tool internal struct
