+++ src/api/tools/__init__.py
@@ -1,14 +1,14 @@
-from .tools import (
-    tools,
-    terminate_virtual_workspace,
-    get_user_id,
-    list_of_regions,
-    workspaces_info,
-    run_sql,
-    choose_organization,
-    organization_info,
-    list_virtual_workspaces,
-)
+# Import from organized structure
+from .tools import tools
+
+# Import individual tools for backward compatibility
+from .user import get_user_id
+from .regions import list_regions as list_of_regions
+from .workspaces import workspaces_info
+from .database import run_sql
+from .organization import choose_organization, organization_info
+from .starter_workspaces import list_virtual_workspaces, terminate_virtual_workspace
+
 from .registery import register_tools
 
 __all__ = [
