+++ src/api/tools/database/__init__.py
@@ -1,0 +1,5 @@
+"""Database tools for SingleStore MCP server."""
+
+from .database import run_sql
+
+__all__ = ["run_sql"]
