+++ src/api/tools/notebooks/__init__.py
@@ -1,0 +1,5 @@
+"""Notebook tools for SingleStore MCP server."""
+
+from .notebooks import create_notebook_file, list_shared_files, check_if_file_exists
+
+__all__ = ["create_notebook_file", "list_shared_files", "check_if_file_exists"]
