+++ src/api/tools/organization/__init__.py
@@ -1,0 +1,5 @@
+"""Organization tools for SingleStore MCP server."""
+
+from .organization import organization_info, choose_organization, set_organization
+
+__all__ = ["organization_info", "choose_organization", "set_organization"]
