+++ src/api/tools/regions/__init__.py
@@ -1,0 +1,5 @@
+"""Regions tools for SingleStore MCP server."""
+
+from .regions import list_regions
+
+__all__ = ["list_regions"]
