+++ src/api/tools/starter_workspaces/__init__.py
@@ -1,0 +1,15 @@
+"""Starter workspaces tools for SingleStore MCP server."""
+
+from .starter_workspaces import (
+    list_virtual_workspaces,
+    create_starter_workspace,
+    terminate_virtual_workspace,
+    create_virtual_workspace,
+)
+
+__all__ = [
+    "list_virtual_workspaces",
+    "create_starter_workspace",
+    "terminate_virtual_workspace",
+    "create_virtual_workspace",
+]
