+++ src/api/tools/user/__init__.py
@@ -1,0 +1,5 @@
+"""User tools for SingleStore MCP server."""
+
+from .user import get_user_id
+
+__all__ = ["get_user_id"]
