+++ src/api/tools/workspaces/__init__.py
@@ -1,0 +1,5 @@
+"""Workspaces tools for SingleStore MCP server."""
+
+from .workspaces import workspaces_info
+
+__all__ = ["workspaces_info"]
